cost-optimization/hpa-config-recommender/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpaconfigrecommender"
version = "0.0.45"
authors = [
    { name="Google LLC", email="ameenahb@google.com" }
]

description = "Workload Recommender Core Library"
readme = "README.md"
requires-python = ">=3.10"


classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]

dependencies = [
    "pandas==2.2.2",
    "pyarrow<19.0.0",
    "numpy==1.26.4",
    "aiohttp>=3.8.0",
    "httpx>=0.28.1",
    "asyncio==3.4.3",
    "requests==2.32.3",
    "google-api-core==2.22.0",
    "google-cloud-asset==3.27.0",
    "google_cloud_bigquery==3.29.0"
]
license = {text = "The Unlicense"}